Refactor GameSetup to a folder
