+++ app/src/main/java/com/bestCatHustlers/sukodublitz/MainActivityPresenter.java
@@ -5,6 +5,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
 import com.bestCatHustlers.sukodublitz.multiplayer.MultiplayerMenuActivity;
+import com.bestCatHustlers.sukodublitz.setup.GameSetupActivity;
 
 
 class MainActivityPresenter {
@@ -14,11 +15,11 @@ class MainActivityPresenter {
         mainActivity = activity;
     }
 
-    protected void goToSinglePlayer(View view) {
+    void goToSinglePlayer(View view) {
         Intent intent = new Intent(mainActivity, GameSetupActivity.class);
         mainActivity.startActivity(intent);
     }
-    protected void goToMultiplayer(View view) {
+    void goToMultiplayer(View view) {
         Intent intent = new Intent(mainActivity, MultiplayerMenuActivity.class);
         mainActivity.startActivity(intent);
     }
