+++ app/src/main/java/com/bestCatHustlers/sukodublitz/game/GamePresenter.java
@@ -1,22 +1,15 @@
 package com.bestCatHustlers.sukodublitz.game;
 
-import android.content.ComponentName;
 import android.content.Intent;
-import android.content.ServiceConnection;
 import android.os.Bundle;
-import android.os.IBinder;
 import android.os.Parcelable;
 import android.os.SystemClock;
-import android.widget.Toast;
 
 import com.bestCatHustlers.sukodublitz.BoardGame;
 import com.bestCatHustlers.sukodublitz.GameAI;
-import com.bestCatHustlers.sukodublitz.GameSetupActivity;
+import com.bestCatHustlers.sukodublitz.setup.GameSetupActivity;
 import com.bestCatHustlers.sukodublitz.Player;
 import com.bestCatHustlers.sukodublitz.R;
-import com.bestCatHustlers.sukodublitz.bluetooth.BluetoothConstants;
-import com.bestCatHustlers.sukodublitz.bluetooth.BluetoothService;
-import com.bestCatHustlers.sukodublitz.utils.ParcelableByteUtil;
 
 import static com.bestCatHustlers.sukodublitz.multiplayer.MultiplayerMenuPresenter.EXTRAS_KEY_IS_MULTI;
