+++ app/src/main/java/com/bestCatHustlers/sukodublitz/multiplayer/MultiplayerMenuActivity.java
@@ -5,7 +5,7 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-import com.bestCatHustlers.sukodublitz.GameSetupActivity;
+import com.bestCatHustlers.sukodublitz.setup.GameSetupActivity;
 import com.bestCatHustlers.sukodublitz.R;
 import com.bestCatHustlers.sukodublitz.join.JoinActivity;
