+++ app/src/main/java/com/bestCatHustlers/sukodublitz/results/ResultsActivity.java
@@ -6,7 +6,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
 
-import com.bestCatHustlers.sukodublitz.GameSetupActivity;
+import com.bestCatHustlers.sukodublitz.setup.GameSetupActivity;
 import com.bestCatHustlers.sukodublitz.MainActivity;
 import com.bestCatHustlers.sukodublitz.R;
 import com.bestCatHustlers.sukodublitz.game.GameBoardView;
