+++ app/src/main/res/layout/activity_game_setup.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_gradient_red_blue"
-    tools:context=".GameSetupActivity">
+    tools:context=".setup.GameSetupActivity">
 
     <Button
         android:id="@+id/back_button"
